monitor/requirements.txt
python-dotenv==1.0.0
requests==2.31.0
aiohttp==3.9.1
aioimaplib==1.0.1
python-telegram-bot==20.7
asyncio==3.4.3
configparser==6.0.0
